refactor: complete runtime handler contract
