+++ agently/builtins/plugins/ModelRequester/OpenAICompatible/plugin.py
@@ -12,42 +12,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import asyncio
-import time
-import yaml
-import json
-from typing import (
-    Any,
-    Literal,
-    AsyncGenerator,
-    TYPE_CHECKING,
-    cast,
-)
 
-from httpx import AsyncClient, ReadError, HTTPStatusError, RequestError, Timeout
-from httpx_sse import aconnect_sse, SSEError
-from stamina import retry
+from __future__ import annotations
 
-from .modules.types import ContentMapping
-from agently.core.AttemptRunner import AttemptRunner, core_attempt_runner_entrypoint
+from typing import TYPE_CHECKING, cast
+
+from httpx import AsyncClient
+
 from agently.types.plugins import ModelRequester
-from agently.types.data import AgentlyRequestData, AttemptDecision, AttemptHandlers, AttemptState, SerializableValue
-from agently.core.AgentExecution import RuntimeStageStallError
 from agently.types.settings import OpenAICompatibleSettings
-from agently.utils import (
-    SettingsNamespace,
-    DataFormatter,
-    DataLocator,
-)
-from agently.utils.ModelPool import resolve_api_key_failover
+from agently.utils import SettingsNamespace
+
+from .modules.credential import OpenAICompatibleCredentialMixin
+from .modules.handlers import OpenAICompatibleHandlersMixin
+from .modules.request_builder import OpenAICompatibleRequestBuilderMixin
+from .modules.response_adapter import OpenAICompatibleResponseAdapterMixin
+from .modules.transport import OpenAICompatibleTransportMixin
 
 if TYPE_CHECKING:
     from agently.core.Prompt import Prompt
     from agently.utils import Settings
-    from agently.types.data import AgentlyResponseGenerator, AgentlyRequestDataDict, ChatMessage
 
 
-class OpenAICompatible(ModelRequester):
+class OpenAICompatible(
+    OpenAICompatibleHandlersMixin,
+    OpenAICompatibleRequestBuilderMixin,
+    OpenAICompatibleTransportMixin,
+    OpenAICompatibleCredentialMixin,
+    OpenAICompatibleResponseAdapterMixin,
+    ModelRequester,
+):
     name = "OpenAICompatible"
     SETTINGS_SCHEMAS = {
         "plugins.ModelRequester.OpenAICompatible": OpenAICompatibleSettings,
@@ -130,751 +124,3 @@ class OpenAICompatible(ModelRequester):
     @staticmethod
     def _on_unregister():
         pass
-
-    def _get_timeout_mode(self) -> Literal["http", "first_token"]:
-        timeout_mode = self.plugin_settings.get("timeout_mode", "first_token")
-        if timeout_mode == "http":
-            return "http"
-        return "first_token"
-
-    def _get_timeout_configs(self) -> dict[str, Any]:
-        return DataFormatter.to_str_key_dict(
-            self.plugin_settings.get(
-                "timeout",
-                {
-                    "connect": 30.0,
-                    "read": 120.0,
-                    "write": 30.0,
-                    "pool": 30.0,
-                },
-            ),
-            default_value={},
-        )
-
-    def _get_http_timeout(self, *, disable_read: bool = False) -> Timeout:
-        timeout_configs = self._get_timeout_configs().copy()
-        if disable_read:
-            timeout_configs["read"] = None
-        return Timeout(**timeout_configs)
-
-    def _get_first_token_timeout_seconds(self) -> float | None:
-        read_timeout = self._get_timeout_configs().get("read")
-        if isinstance(read_timeout, (int, float)) and read_timeout > 0:
-            return float(read_timeout)
-        return None
-
-    def _get_stream_idle_timeout_seconds(self) -> float | None:
-        raw_timeout = self.plugin_settings.get("stream_idle_timeout", None)
-        if raw_timeout is None or raw_timeout == -1 or raw_timeout == "-1":
-            return None
-        if isinstance(raw_timeout, bool) or not isinstance(raw_timeout, (int, float, str)):
-            return None
-        try:
-            timeout = float(raw_timeout)
-        except (TypeError, ValueError):
-            return None
-        return timeout if timeout > 0 else None
-
-    def _should_use_first_token_timeout(self, request_data: "AgentlyRequestData") -> bool:
-        return (
-            self._get_timeout_mode() == "first_token"
-            and self.model_type in ("chat", "completions")
-            and bool(request_data.stream)
-        )
-
-    def _build_stream_stall_error(
-        self,
-        *,
-        stage: str,
-        timeout_seconds: float,
-        message: str,
-    ) -> RuntimeStageStallError:
-        return RuntimeStageStallError(
-            message,
-            stage=stage,
-            status="stalled",
-            idle_seconds=timeout_seconds,
-            timeout_seconds=timeout_seconds,
-            provider=self.name,
-            model=cast(str | None, self.plugin_settings.get("model", None)),
-        )
-
-    async def _aiter_with_first_token_timeout(
-        self,
-        generator: AsyncGenerator[Any, None],
-        *,
-        timeout_seconds: float | None,
-    ) -> AsyncGenerator[Any, None]:
-        if timeout_seconds is None:
-            async for item in generator:
-                yield item
-            return
-
-        try:
-            first_item = await asyncio.wait_for(anext(generator), timeout=timeout_seconds)
-        except asyncio.TimeoutError as e:
-            await generator.aclose()
-            raise self._build_stream_stall_error(
-                stage="response_first_event",
-                timeout_seconds=timeout_seconds,
-                message=f"First token timeout after { timeout_seconds } seconds.",
-            ) from e
-
-        yield first_item
-        async for item in generator:
-            yield item
-
-    async def _aiter_with_stream_idle_timeout(
-        self,
-        generator: AsyncGenerator[Any, None],
-        *,
-        timeout_seconds: float | None,
-    ) -> AsyncGenerator[Any, None]:
-        if timeout_seconds is None:
-            async for item in generator:
-                yield item
-            return
-
-        try:
-            first_item = await anext(generator)
-        except StopAsyncIteration:
-            return
-
-        yield first_item
-        while True:
-            try:
-                item = await asyncio.wait_for(anext(generator), timeout=timeout_seconds)
-            except StopAsyncIteration:
-                return
-            except asyncio.TimeoutError as e:
-                await generator.aclose()
-                raise self._build_stream_stall_error(
-                    stage="response_stream",
-                    timeout_seconds=timeout_seconds,
-                    message=f"Stream idle timeout after { timeout_seconds } seconds.",
-                ) from e
-            yield item
-
-    def generate_request_data(self) -> "AgentlyRequestData":
-        agently_request_dict: AgentlyRequestDataDict = {
-            "client_options": {},
-            "headers": {},
-            "data": {},
-            "request_options": {},
-            "request_url": "",
-        }
-        # main data
-        match self.model_type:
-            case "chat":
-                request_data = {
-                    "messages": self.prompt.to_messages(
-                        rich_content=bool(
-                            self.plugin_settings.get(
-                                "rich_content",
-                                False,
-                            )
-                        ),
-                        strict_role_orders=bool(
-                            self.plugin_settings.get(
-                                "strict_role_orders",
-                                False,
-                            ),
-                        ),
-                    )
-                }
-            case "completions":
-                request_data = {"prompt": self.prompt.to_text()}
-            case "embeddings":
-                sanitized_input = DataFormatter.sanitize(self.prompt["input"])
-                if isinstance(sanitized_input, list):
-                    request_data = {
-                        "input": [
-                            (
-                                str(item)
-                                if isinstance(item, (str, int, float, bool)) or item is None
-                                else yaml.safe_dump(item)
-                            )
-                            for item in sanitized_input
-                        ],
-                    }
-                else:
-                    request_data = {
-                        "input": (
-                            str(sanitized_input)
-                            if isinstance(sanitized_input, (str, int, float, bool)) or sanitized_input is None
-                            else yaml.safe_dump(sanitized_input)
-                        )
-                    }
-            case _:
-                raise TypeError(
-                    f"Plugin Name: { self.name }\n" f"Error: Cannot support model type: '{ self.model_type }'"
-                )
-                request_data = {}
-        ## set
-        agently_request_dict["data"] = request_data
-
-        # headers
-        headers: dict[str, str] = DataFormatter.to_str_key_dict(
-            self.plugin_settings.get("headers"),
-            value_format="str",
-            default_value={},
-        )
-        headers.update({"Connection": "close"})
-        ## set
-        agently_request_dict["headers"] = headers
-
-        # client options
-        client_options = DataFormatter.to_str_key_dict(self.plugin_settings.get("client_options"), default_value={})
-        client_options.setdefault("trust_env", False)
-        ## proxy
-        proxy = self.plugin_settings.get("proxy", None)
-        if proxy:
-            client_options.update({"proxy": proxy})
-        ## timeout
-        timeout = self._get_http_timeout()
-        client_options.update({"timeout": timeout})
-        ## set
-        agently_request_dict["client_options"] = client_options
-
-        # request_options
-        legacy_options = DataFormatter.to_str_key_dict(
-            self.plugin_settings.get("options"),
-            value_format="serializable",
-            default_value={},
-        )
-        request_options = DataFormatter.to_str_key_dict(
-            self.plugin_settings.get("request_options"),
-            value_format="serializable",
-            default_value={},
-        )
-        # Backward compatibility for older examples/configs that still use
-        # plugin-root `options` for default request-body parameters.
-        request_options = {**legacy_options, **request_options}
-        request_options_in_prompt = self.prompt.get("options", {})
-        if request_options_in_prompt:
-            request_options.update(request_options_in_prompt)
-            request_options = DataFormatter.to_str_key_dict(
-                request_options,
-                value_format="serializable",
-                default_value={},
-            )
-        ## !: ensure model
-        request_options.update(
-            {
-                "model": self.plugin_settings.get(
-                    "model",
-                    DataFormatter.to_str_key_dict(
-                        self.plugin_settings.get("default_model"),
-                        value_format="serializable",
-                        default_key=self.model_type,
-                    )[self.model_type],
-                )
-            }
-        )
-        ## !: ensure stream
-        is_stream = self.plugin_settings.get("stream")
-        if is_stream is None:
-            if self.model_type == "embeddings":
-                is_stream = False
-            else:
-                is_stream = True
-        request_options.update({"stream": is_stream})
-        ## set
-        agently_request_dict["request_options"] = request_options
-
-        # request url
-        ## get full url
-        full_url = self.plugin_settings.get("full_url")
-        ## get base url
-        base_url = str(self.plugin_settings.get("base_url"))
-        base_url = base_url[:-1] if base_url[-1] == "/" else base_url
-        ## get path mapping
-        path_mapping = DataFormatter.to_str_key_dict(
-            self.plugin_settings.get("path_mapping"),
-            value_format="str",
-            default_value={},
-        )
-        path_mapping = {k: v if v[0] == "/" else f"/{ v }" for k, v in path_mapping.items()}
-        ## set
-        if isinstance(full_url, str):
-            request_url = full_url
-        else:
-            request_url = f"{ base_url }{ path_mapping[self.model_type] }"
-        agently_request_dict["request_url"] = request_url
-
-        return AgentlyRequestData(**agently_request_dict)
-
-    async def _aiter_sse_with_retry(
-        self,
-        client: AsyncClient,
-        method: str,
-        url: str,
-        *,
-        headers: dict[str, Any],
-        json: "SerializableValue",
-    ):
-        last_event_id = ""
-        reconnection_delay = 0.0
-
-        @retry(on=ReadError)
-        async def _aiter_sse():
-            nonlocal last_event_id, reconnection_delay
-            time.sleep(reconnection_delay)
-            headers.update({"Accept": "text/event-stream"})
-            if last_event_id:
-                headers.update({"Last-Event-ID": last_event_id})
-
-            async with aconnect_sse(client, method, url, headers=headers, json=json) as event_source:
-                try:
-                    async for sse in event_source.aiter_sse():
-                        last_event_id = sse.id
-                        if sse.retry is not None:
-                            reconnection_delay = sse.retry / 1000
-                        yield sse
-                except GeneratorExit:
-                    pass
-
-        return _aiter_sse()
-
-    def _build_headers_with_auth(self, request_data: "AgentlyRequestData") -> dict[str, Any]:
-        auth = DataFormatter.to_str_key_dict(
-            self.plugin_settings.get("auth", "None"),
-            value_format="serializable",
-            default_key="api_key",
-        )
-        api_key = self.plugin_settings.get("api_key", None)
-        auth_api_key = auth.get("api_key", "None")
-        if api_key is not None and auth_api_key == "None":
-            auth["api_key"] = str(api_key)
-            auth_api_key = auth["api_key"]
-        headers_with_auth = request_data.headers.copy()
-        if "headers" in auth and isinstance(auth["headers"], dict):
-            headers_with_auth.update(
-                DataFormatter.to_str_key_dict(
-                    auth["headers"],
-                    value_format="str",
-                    default_value={},
-                )
-            )
-        if "body" in auth and isinstance(auth["body"], dict):
-            request_data.data.update(**auth["body"])
-        if auth_api_key != "None" and "Authorization" not in headers_with_auth:
-            headers_with_auth["Authorization"] = f"Bearer { auth_api_key }"
-        return headers_with_auth
-
-    def _build_failover_headers(
-        self,
-        request_data: "AgentlyRequestData",
-        *,
-        error: Any,
-        status_code: int | None,
-        response_text: str | None,
-        full_request_data: dict[str, Any],
-        stream_started: bool,
-    ) -> dict[str, Any] | None:
-        decision = resolve_api_key_failover(
-            self.plugin_settings,
-            error=error,
-            status_code=status_code,
-            response_text=response_text,
-            request_data=full_request_data,
-            provider=self.name,
-            stream_started=stream_started,
-        )
-        if not decision.retry:
-            return None
-        return self._build_headers_with_auth(request_data)
-
-    def _build_full_request_data(self, request_data: "AgentlyRequestData") -> dict[str, Any]:
-        full_request_data = DataFormatter.to_str_key_dict(
-            request_data.data,
-            value_format="serializable",
-            default_value={},
-        )
-        full_request_data.update(request_data.request_options)
-        return full_request_data
-
-    def build_request_handlers(self, request_data: "AgentlyRequestData") -> AttemptHandlers:
-        async def execute(_state: AttemptState) -> AsyncGenerator[tuple[str, Any], None]:
-            async for item in self._request_model_legacy(request_data):
-                yield item
-
-        async def handle_error(error: BaseException, _state: AttemptState) -> AttemptDecision:
-            return AttemptDecision.yield_error(error)
-
-        return AttemptHandlers(execute=execute, handle_error=handle_error)
-
-    @core_attempt_runner_entrypoint
-    async def request_model(self, request_data: "AgentlyRequestData") -> AsyncGenerator[tuple[str, Any], None]:
-        runner = AttemptRunner(self.build_request_handlers(request_data))
-        async for item in runner.run_stream():
-            yield item
-
-    async def _request_model_legacy(self, request_data: "AgentlyRequestData") -> AsyncGenerator[tuple[str, Any], None]:
-        headers_with_auth = self._build_headers_with_auth(request_data)
-
-        # request
-        # stream request
-        if self.model_type in ("chat", "completions") and request_data.stream:
-            client_options = request_data.client_options.copy()
-            if self._should_use_first_token_timeout(request_data):
-                client_options.update({"timeout": self._get_http_timeout(disable_read=True)})
-
-            async with AsyncClient(**client_options) as client:
-                client.headers.update(headers_with_auth)
-                full_request_data = DataFormatter.to_str_key_dict(
-                    request_data.data,
-                    value_format="serializable",
-                    default_value={},
-                )
-                full_request_data.update(request_data.request_options)
-                stream_started = False
-                while True:
-                    try:
-                        has_done = False
-                        sse_generator = await self._aiter_sse_with_retry(
-                            client,
-                            "POST",
-                            request_data.request_url,
-                            json=full_request_data,
-                            headers=headers_with_auth,
-                        )
-                        if self._should_use_first_token_timeout(request_data):
-                            sse_generator = self._aiter_with_first_token_timeout(
-                                sse_generator,
-                                timeout_seconds=self._get_first_token_timeout_seconds(),
-                            )
-                        sse_generator = self._aiter_with_stream_idle_timeout(
-                            sse_generator,
-                            timeout_seconds=self._get_stream_idle_timeout_seconds(),
-                        )
-                        async for sse in sse_generator:
-                            stream_started = True
-                            yield sse.event, sse.data
-                            if sse.data.strip() == "[DONE]":
-                                has_done = True
-                        if not has_done:
-                            yield "message", "[DONE]"
-                        break
-                    except SSEError as e:
-                        response = await client.post(
-                            request_data.request_url,
-                            json=full_request_data,
-                            headers=headers_with_auth,
-                        )
-                        if response.status_code >= 400:
-                            e = RequestError(
-                                f"Status Code: { response.status_code }\n"
-                                f"Detail: { response.text }\n"
-                                f"Request Data: {full_request_data}"
-                            )
-                            failover_headers = self._build_failover_headers(
-                                request_data,
-                                error=e,
-                                status_code=response.status_code,
-                                response_text=response.text,
-                                full_request_data=full_request_data,
-                                stream_started=stream_started,
-                            )
-                            if failover_headers is not None:
-                                headers_with_auth = failover_headers
-                                client.headers.update(headers_with_auth)
-                                continue
-                            yield "error", e
-                        else:
-                            content_type = response.headers.get("Content-Type", "")
-                            if content_type.startswith("application/json"):
-                                try:
-                                    error_json = response.json()
-                                except Exception:
-                                    error_json = await response.aread()
-                                    error_json = json.loads(error_json.decode())
-                                error = error_json["error"]
-                                error_title = f"{ error['code'] if 'code' in error else 'unknown_code' } - { error['type'] if 'type' in error else 'unknown_type' }"
-                                error_detail = error["message"] if "message" in error else ""
-                                yield "error", error_detail
-                            else:
-                                yield "error", e
-                        break
-                    except HTTPStatusError as e:
-                        failover_headers = self._build_failover_headers(
-                            request_data,
-                            error=e,
-                            status_code=e.response.status_code,
-                            response_text=e.response.text,
-                            full_request_data=full_request_data,
-                            stream_started=stream_started,
-                        )
-                        if failover_headers is not None:
-                            headers_with_auth = failover_headers
-                            client.headers.update(headers_with_auth)
-                            continue
-                        yield "error", e
-                        break
-                    except TimeoutError as e:
-                        failover_headers = self._build_failover_headers(
-                            request_data,
-                            error=e,
-                            status_code=None,
-                            response_text=None,
-                            full_request_data=full_request_data,
-                            stream_started=stream_started,
-                        )
-                        if failover_headers is not None:
-                            headers_with_auth = failover_headers
-                            client.headers.update(headers_with_auth)
-                            continue
-                        yield "error", e
-                        break
-                    except RequestError as e:
-                        failover_headers = self._build_failover_headers(
-                            request_data,
-                            error=e,
-                            status_code=None,
-                            response_text=None,
-                            full_request_data=full_request_data,
-                            stream_started=stream_started,
-                        )
-                        if failover_headers is not None:
-                            headers_with_auth = failover_headers
-                            client.headers.update(headers_with_auth)
-                            continue
-                        yield "error", e
-                        break
-                    except Exception as e:
-                        yield "error", e
-                        break
-        # normal request
-        else:
-            async with AsyncClient(**request_data.client_options) as client:
-                client.headers.update(headers_with_auth)
-                full_request_data = DataFormatter.to_str_key_dict(
-                    request_data.data,
-                    value_format="serializable",
-                    default_value={},
-                )
-                full_request_data.update(request_data.request_options)
-                while True:
-                    try:
-                        response = await client.post(
-                            request_data.request_url,
-                            json=full_request_data,
-                            headers=headers_with_auth,
-                        )
-                        if response.status_code >= 400:
-                            e = RequestError(
-                                f"Status Code: { response.status_code }\n"
-                                f"Detail: { response.text }\n"
-                                f"Request Data: {full_request_data}"
-                            )
-                            failover_headers = self._build_failover_headers(
-                                request_data,
-                                error=e,
-                                status_code=response.status_code,
-                                response_text=response.text,
-                                full_request_data=full_request_data,
-                                stream_started=False,
-                            )
-                            if failover_headers is not None:
-                                headers_with_auth = failover_headers
-                                client.headers.update(headers_with_auth)
-                                continue
-                            yield "error", e
-                        else:
-                            yield "message", response.content.decode()
-                            yield "message", "[DONE]"
-                        break
-                    except HTTPStatusError as e:
-                        failover_headers = self._build_failover_headers(
-                            request_data,
-                            error=e,
-                            status_code=e.response.status_code,
-                            response_text=e.response.text,
-                            full_request_data=full_request_data,
-                            stream_started=False,
-                        )
-                        if failover_headers is not None:
-                            headers_with_auth = failover_headers
-                            client.headers.update(headers_with_auth)
-                            continue
-                        yield "error", e
-                        break
-                    except RequestError as e:
-                        failover_headers = self._build_failover_headers(
-                            request_data,
-                            error=e,
-                            status_code=None,
-                            response_text=None,
-                            full_request_data=full_request_data,
-                            stream_started=False,
-                        )
-                        if failover_headers is not None:
-                            headers_with_auth = failover_headers
-                            client.headers.update(headers_with_auth)
-                            continue
-                        yield "error", e
-                        break
-                    except Exception as e:
-                        yield "error", e
-                        break
-
-    async def broadcast_response(self, response_generator: AsyncGenerator) -> "AgentlyResponseGenerator":
-        meta = {}
-        message_record = {}
-        reasoning_buffer = ""
-        content_buffer = ""
-
-        content_mapping = cast(
-            ContentMapping,
-            DataFormatter.to_str_key_dict(
-                self.plugin_settings.get("content_mapping"),
-                value_format="serializable",
-            ),
-        )
-        id_mapping = content_mapping["id"]
-        role_mapping = content_mapping["role"]
-        reasoning_mapping = content_mapping["reasoning"]
-        delta_mapping = content_mapping["delta"]
-        tool_calls_mapping = content_mapping["tool_calls"]
-        done_mapping = content_mapping["done"]
-        usage_mapping = content_mapping["usage"]
-        finish_reason_mapping = content_mapping["finish_reason"]
-        extra_delta_mapping = content_mapping["extra_delta"]
-        extra_done_mapping = content_mapping["extra_done"]
-        yield_extra_content_separately = self.plugin_settings.get("yield_extra_content_separately", True)
-
-        content_mapping_style = str(self.plugin_settings.get("content_mapping_style"))
-        if content_mapping_style not in ("dot", "slash"):
-            content_mapping_style = "dot"
-
-        async for event, message in response_generator:
-            if event == "error":
-                yield "error", message
-            elif message != "[DONE]":
-                yield "original_delta", message
-                loaded_message = json.loads(message)
-                message_record = loaded_message.copy()
-                if "id" not in meta and id_mapping:
-                    _id = DataLocator.locate_path_in_dict(
-                        loaded_message,
-                        id_mapping,
-                        style=content_mapping_style,
-                    )
-                    if _id:
-                        meta.update({"id": _id})
-                if "role" not in meta and role_mapping:
-                    role = DataLocator.locate_path_in_dict(
-                        loaded_message,
-                        role_mapping,
-                        style=content_mapping_style,
-                        default="assistant",
-                    )
-                    if role:
-                        meta.update({"role": role})
-                if reasoning_mapping:
-                    reasoning = DataLocator.locate_path_in_dict(
-                        loaded_message,
-                        reasoning_mapping,
-                        style=content_mapping_style,
-                    )
-                    if reasoning:
-                        reasoning_buffer += str(reasoning)
-                        yield "reasoning_delta", reasoning
-                if delta_mapping:
-                    delta = DataLocator.locate_path_in_dict(
-                        loaded_message,
-                        delta_mapping,
-                        style=content_mapping_style,
-                    )
-                    if delta:
-                        content_buffer += str(delta)
-                        yield "delta", delta
-                if tool_calls_mapping:
-                    tool_calls = DataLocator.locate_path_in_dict(
-                        loaded_message,
-                        tool_calls_mapping,
-                        style=content_mapping_style,
-                    )
-                    if tool_calls:
-                        yield "tool_calls", tool_calls
-                if extra_delta_mapping:
-                    for extra_key, extra_path in extra_delta_mapping.items():
-                        extra_value = DataLocator.locate_path_in_dict(
-                            loaded_message,
-                            extra_path,
-                            style=content_mapping_style,
-                        )
-                        if extra_value:
-                            yield "extra", {extra_key: extra_value}
-                            if yield_extra_content_separately:
-                                yield extra_key, extra_value  # type: ignore
-            else:
-                done_content = None
-                if self.model_type == "embeddings" and done_mapping is None:
-                    done_mapping = "data"
-                    content_mapping_style = "dot"
-                if done_mapping:
-                    done_content = DataLocator.locate_path_in_dict(
-                        message_record,
-                        done_mapping,
-                        style=content_mapping_style,
-                    )
-                if done_content:
-                    yield "done", done_content
-                else:
-                    yield "done", content_buffer
-                reasoning_content = None
-                if reasoning_mapping:
-                    reasoning_content = DataLocator.locate_path_in_dict(
-                        message_record,
-                        reasoning_mapping,
-                        style=content_mapping_style,
-                    )
-                if reasoning_content:
-                    yield "reasoning_done", reasoning_content
-                else:
-                    yield "reasoning_done", reasoning_buffer
-                match self.model_type:
-                    case "embeddings":
-                        yield "original_done", message_record
-                    case "_":
-                        done_message = message_record
-                        if "message" not in done_message["choices"][0]:
-                            done_message["choices"][0].update({"message": {}})
-                        done_message["choices"][0]["message"].update(
-                            {
-                                "role": meta["role"] if "role" in meta else "assistant",
-                                "content": done_content if done_content else content_buffer,
-                            }
-                        )
-                        yield "original_done", done_message
-                if finish_reason_mapping:
-                    meta.update(
-                        {
-                            "finish_reason": DataLocator.locate_path_in_dict(
-                                message_record,
-                                finish_reason_mapping,
-                                style=content_mapping_style,
-                            )
-                        }
-                    )
-                if usage_mapping:
-                    meta.update(
-                        {
-                            "usage": DataLocator.locate_path_in_dict(
-                                message_record,
-                                usage_mapping,
-                                style=content_mapping_style,
-                            )
-                        }
-                    )
-                yield "meta", meta
-                if extra_done_mapping:
-                    for extra_key, extra_path in extra_done_mapping.items():
-                        extra_value = DataLocator.locate_path_in_dict(
-                            message_record,
-                            extra_path,
-                            style=content_mapping_style,
-                        )
-                        if extra_value:
-                            yield "extra", {extra_key: extra_value}
