+++ compatibility/in-development.json
@@ -18,7 +18,12 @@
         "runtime_event_ownership": {
           "official_event_producer": "core",
           "plugin_contract": "plugins return observations/errors/decisions; core maps them to official RuntimeEvent records",
-          "builtin_direct_emitters_for_official_events": false
+          "builtin_direct_emitters_for_official_events": false,
+          "agent_execution_stream_owner": "agently.core.AgentExecution.AgentExecutionStream"
+        },
+        "runtime_naming": {
+          "agent_turn": "run_kind for one Agent-facing turn",
+          "attempt_index": "model-request retry attempt metadata; not an agent turn counter"
         },
         "agent_execution_limits": [
           "max_seconds",
