+++ tests/test_compatibility_registry.py
@@ -70,6 +70,16 @@ def test_in_development_manifest_is_registered_and_protocol_compatible():
         "compatibility_input_type": "ObservationEvent",
     }
     assert in_development["companions"]["devtools"]["runtime_control"] == {
+        "runtime_event_ownership": {
+            "official_event_producer": "core",
+            "plugin_contract": "plugins return observations/errors/decisions; core maps them to official RuntimeEvent records",
+            "builtin_direct_emitters_for_official_events": False,
+            "agent_execution_stream_owner": "agently.core.AgentExecution.AgentExecutionStream",
+        },
+        "runtime_naming": {
+            "agent_turn": "run_kind for one Agent-facing turn",
+            "attempt_index": "model-request retry attempt metadata; not an agent turn counter",
+        },
         "agent_execution_limits": ["max_seconds", "max_no_progress_seconds"],
         "provider_stream_idle_timeout": [
             "OpenAICompatible.stream_idle_timeout",
